fix(tui): bound and correct the full-screen transcript scroll
The full-screen renderer's scroll path had three defects in long sessions,
addressed together because they share the same code:

1. Per-frame full-document rewrap (perf). `full_transcript_lines` markdown/
   syntax-highlighted and word-wrapped the *entire* history every frame just
   to show one viewport — O(session) work per frame. Because `App::lines` only
   grows at the tail between resets, wrapping is now memoized in a
   `TranscriptWrapCache`: a width- and generation-keyed cache re-wraps only the
   newly-appended lines. `full_transcript_lines` becomes a resumable
   `append_transcript_range` the cache drives incrementally.

2. u16 overflow (correctness). `ScrollState::offset`, `content_height`, and the
   full-screen scroll metrics were all `u16`. At 65,536 wrapped rows the content
   height wrapped to ~0, collapsing stick-to-bottom's `max_offset` and silently
   snapping the transcript to the top. Widened the scroll dimensions to `usize`
   throughout `tuika::Scroll`/`ScrollState` and the host metrics.

3. Unbounded memory. `App::lines` was only ever pushed/extended/cleared, so both
   renderers retained the whole session in RAM. `trim_transcript` now caps
   retained history at `MAX_RETAINED_TRANSCRIPT_LINES`, draining the front once
   per frame; inline mode only drops lines already flushed to native scrollback,
   full-screen mode bounds its in-app scrollback to the window. The drain shifts
   the flush cursor and banner length and invalidates the wrap cache.

Tests: wrap-cache-vs-full-rebuild equality (incremental and width-change),
usize offset survives a >u16::MAX-row transcript, and the trim cap for both
render modes.

+++ crates/tuika/examples/demo.rs
@@ -707,8 +707,8 @@ fn scene_scroll(frame: u64, theme: &Theme) -> Element {
             ))
         })
         .collect();
-    let viewport_h = 8u16;
-    let content_h = lines.len() as u16;
+    let viewport_h = 8usize;
+    let content_h = lines.len();
     let mut state = ScrollState::new();
     let reach = tuika::anim::ping_pong(frame, 200);
     let steps = (reach * (content_h.saturating_sub(viewport_h) as f32 / 3.0 + 1.0)) as u32;
